Convert PixelMap to use SQL
place cooldowns are still needed

+++ Source/Commands/PixelCommands.cs
@@ -76,7 +76,7 @@ namespace DiscordBotRewrite.Commands {
 
             DiscordSelectComponent colorSelectComponent = new DiscordSelectComponent("color", "Select color to place:", colorOptions);
 
-            PixelMap map = Bot.Modules.Pixel.GetPixelMap(ctx.Guild.Id);
+            PixelMap map = Bot.Modules.Pixel.GetPixelMap((long)ctx.Guild.Id);
 
             int curX = (int)Math.Clamp(x, 0, map.Width - 1);
             int curY = (int)Math.Clamp(y, 0, map.Height - 1);
@@ -188,7 +188,7 @@ namespace DiscordBotRewrite.Commands {
             [Option("x", "new x size")] long x,
             [Option("y", "new y size")] long y) {
 
-            Bot.Modules.Pixel.ResizeMap(ctx.Guild.Id, (int)x, (int)y);
+            Bot.Modules.Pixel.ResizeMap((long)ctx.Guild.Id, (int)x, (int)y);
             await ctx.CreateResponseAsync($"Resized Canvas to ({x},{y})!");
         }
         #endregion
@@ -198,7 +198,7 @@ namespace DiscordBotRewrite.Commands {
         [RequirePermissions(Permissions.Administrator)]
         public async Task Cooldown(InteractionContext ctx, [Option("duration", "Time in seconds")] long duration) {
             duration = System.Math.Max(0, duration);
-            Bot.Modules.Pixel.SetCooldown(ctx.Guild.Id, (uint)duration);
+            Bot.Modules.Pixel.SetCooldown((long)ctx.Guild.Id, (uint)duration);
             await ctx.CreateResponseAsync($"Set cooldown to {duration} seconds");
         }
         #endregion
